majo change in design concepts that have been necessary to allow the creation of different integration point lists, i.e. the quadratures and the lagrange points.
The already exsisting quadratures should work exactly as before.



git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@1622 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/quadrature/caching/pointprovider.cc
@@ -20,7 +20,7 @@ namespace Dune {
                        (quad.id(),
                         GlobalPointVectorType(quad.nop()))
                        ).first;
-      for (int i = 0; i < quad.nop(); ++i) {
+      for (size_t i = 0; i < quad.nop(); ++i) {
         it->second[i] = quad.point(i);
       }
       StorageInterface<dim>::registerQuadratureToStorages(quad);
@@ -51,7 +51,7 @@ namespace Dune {
     MapperIteratorType it = mappers_.find(quad.id());
     if (it == mappers_.end()) {
       std::vector<LocalPointType> pts(quad.nop());
-      for (int i = 0; i < quad.nop(); ++i) {
+      for (size_t i = 0; i < quad.nop(); ++i) {
         pts[i] = quad.point(i);
       }
       it = addEntry(quad, pts, elementGeo);
